Implement A* search algorithm
Dealt with the addition of a method that contains the A* search
algorithm. Node and Vector2i classes were created to be used as tools for
the algorithm. Additionally, the collision system was fixed but could
still use some work because mob collision is calculated in the exact
center of the mob, so they clip through the wall slightly. Also, deleted
the unused SpriteRotation class. The rotation of sprites is still a task
that must be tackled later. Finally and of special note, the first
lambda expression in the project was used to override the compare method
of the functional interface Comparator.