Create button classes and button listener interfaces
Migrated the static sprites and spritesheets in SpriteCollection and SpriteSheetCollection over to a new interface called Sprites, changing TileCollection into Tiles (interface) as well. Cleaned up a bunch of imports. Created a UIButtonListener and UIActionListener interface (deviation from TheCherno). Create instance of action interface as a parameter to the UIButton in player class using a lambda expression.